# BileMo/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BileMo",
        "description": "This API provides to BileMo referenced user to access its products and manage their customers",
        "version": "1.0.0"
    },
    "paths": {
        "/api/customers/users/{user_id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Provides the list of customer resources for a specific user.",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Your user's ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The collection page number",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer list"
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Provides one product resource for a specific user.",
                "operationId": "App\\Controller\\CustomerController::item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "customer's Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The resource doesn't exist"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Provide the list of all the product resources.",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The collection page number",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product#list"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Provides one product resource.",
                "operationId": "App\\Controller\\ProductController::details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product's Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The resource doesn't exist"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "properties": {
                    "id": {
                        "title": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "title": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "reference": {
                        "type": "string"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "storageCapacity": {
                        "type": "string"
                    },
                    "operatingSystem": {
                        "type": "string"
                    },
                    "screenSize": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        }
    }
}